Move code common used for benchmarks and tests into Utilities.jl
