+++ test/runtests.jl
@@ -1,8 +1,10 @@
 using GenomicFeatures
 using Test
 
-using Distributions
 import Random
+
+include(joinpath(@__DIR__, "Utilities.jl"))
+import .Utilities: random_intervals
 
 
 # Test that an array of intervals is well ordered
@@ -13,27 +15,6 @@ function is_all_ordered(intervals::Vector{I}) where I <: Interval
         end
     end
     return true
-end
-
-# Generate an array of n random Interval{Int} object. With sequence names
-# samples from seqnames, and intervals drawn to lie in [1, maxpos].
-function random_intervals(seqnames, maxpos::Int, n::Int)
-    seq_dist = Categorical(length(seqnames))
-    strand_dist = Categorical(2)
-    length_dist = Normal(1000, 1000)
-    intervals = Vector{Interval{Int}}(undef, n)
-    for i in 1:n
-        intlen = maxpos
-        while intlen >= maxpos || intlen <= 0
-            intlen = ceil(Int, rand(length_dist))
-        end
-        first = rand(1:maxpos-intlen)
-        last = first + intlen - 1
-        strand = rand(strand_dist) == 1 ? STRAND_POS : STRAND_NEG
-        intervals[i] = Interval{Int}(seqnames[rand(seq_dist)],
-                                     first, last, strand, i)
-    end
-    return intervals
 end
 
 # A simple interval intersection implementation to test against.
